added logs to services

+++ Backend/artwork-space-master/src/main/java/com/system/artworkspace/artwork/ArtworkServiceImpl.java
@@ -1,6 +1,9 @@
 package com.system.artworkspace.artwork;
 
+import com.system.artworkspace.ArtworkSpaceApplication;
 import jakarta.persistence.EntityNotFoundException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
@@ -12,6 +15,7 @@ import java.util.Optional;
 
 @Service
 public class ArtworkServiceImpl implements ArtworkService{
+    static final Logger logger = LoggerFactory.getLogger(ArtworkSpaceApplication.class);
 
     private final ArtworkRepository repository;
 
@@ -22,14 +26,27 @@ public class ArtworkServiceImpl implements ArtworkService{
 
     @Override
     public Artwork addArtwork(Artwork artwork) {
+        logger.info("Adding artwork with ID: {}", artwork.getId());
+
         repository.save(artwork);
+        logger.info("Artwork added successfully.");
+
         return artwork;
     }
 
     @Override
     public void deleteArtwork(Long artworkId) {
-repository.deleteById(artworkId);
+        Optional<Artwork> optionalArtwork = repository.findById(artworkId);
+
+        if (optionalArtwork.isPresent()) {
+            repository.deleteById(artworkId);
+            logger.info("Artwork deleted with ID: {}", artworkId);
+        } else {
+            logger.warn("Artwork not found for deletion with ID: {}", artworkId);
+            throw new EntityNotFoundException("Artwork not found with ID: " + artworkId);
+        }
     }
+
 
     @Override
     public Artwork findArtworkById(Long id) {
@@ -45,6 +62,7 @@ repository.deleteById(artworkId);
             Artwork existingArtwork= optionalArtwork.get();
             existingArtwork.setTitle(title);
             repository.save(existingArtwork);
+            logger.info("Title updated for artwork with ID: {}", artwork.getId());
         } else {
             throw new EntityNotFoundException("Artwork not found with ID: " + artwork.getId());
         }
@@ -55,9 +73,10 @@ repository.deleteById(artworkId);
         Optional<Artwork> optionalArtwork = repository.findById(artwork.getId());
 
         if (optionalArtwork.isPresent()) {
-            Artwork existingArtwork= optionalArtwork.get();
+            Artwork existingArtwork = optionalArtwork.get();
             existingArtwork.setDescription(description);
             repository.save(existingArtwork);
+            logger.info("Description updated for artwork with ID: {}", artwork.getId());
         } else {
             throw new EntityNotFoundException("Artwork not found with ID: " + artwork.getId());
         }
@@ -68,9 +87,10 @@ repository.deleteById(artworkId);
         Optional<Artwork> optionalArtwork = repository.findById(artwork.getId());
 
         if (optionalArtwork.isPresent()) {
-            Artwork existingArtwork= optionalArtwork.get();
+            Artwork existingArtwork = optionalArtwork.get();
             existingArtwork.setTechnique(technique);
             repository.save(existingArtwork);
+            logger.info("Technique updated for artwork with ID: {}", artwork.getId());
         } else {
             throw new EntityNotFoundException("Artwork not found with ID: " + artwork.getId());
         }
@@ -81,9 +101,10 @@ repository.deleteById(artworkId);
         Optional<Artwork> optionalArtwork = repository.findById(artwork.getId());
 
         if (optionalArtwork.isPresent()) {
-            Artwork existingArtwork= optionalArtwork.get();
+            Artwork existingArtwork = optionalArtwork.get();
             existingArtwork.setWidth(width);
             repository.save(existingArtwork);
+            logger.info("Width updated for artwork with ID: {}", artwork.getId());
         } else {
             throw new EntityNotFoundException("Artwork not found with ID: " + artwork.getId());
         }
@@ -94,9 +115,10 @@ repository.deleteById(artworkId);
         Optional<Artwork> optionalArtwork = repository.findById(artwork.getId());
 
         if (optionalArtwork.isPresent()) {
-            Artwork existingArtwork= optionalArtwork.get();
+            Artwork existingArtwork = optionalArtwork.get();
             existingArtwork.setHeight(height);
             repository.save(existingArtwork);
+            logger.info("Height updated for artwork with ID: {}", artwork.getId());
         } else {
             throw new EntityNotFoundException("Artwork not found with ID: " + artwork.getId());
         }
@@ -107,9 +129,10 @@ repository.deleteById(artworkId);
         Optional<Artwork> optionalArtwork = repository.findById(artwork.getId());
 
         if (optionalArtwork.isPresent()) {
-            Artwork existingArtwork= optionalArtwork.get();
+            Artwork existingArtwork = optionalArtwork.get();
             existingArtwork.setImageURL(url);
             repository.save(existingArtwork);
+            logger.info("Image URL updated for artwork with ID: {}", artwork.getId());
         } else {
             throw new EntityNotFoundException("Artwork not found with ID: " + artwork.getId());
         }
@@ -120,13 +143,15 @@ repository.deleteById(artworkId);
         Optional<Artwork> optionalArtwork = repository.findById(artwork.getId());
 
         if (optionalArtwork.isPresent()) {
-            Artwork existingArtwork= optionalArtwork.get();
+            Artwork existingArtwork = optionalArtwork.get();
             existingArtwork.setImageSize(size);
             repository.save(existingArtwork);
+            logger.info("Image Size updated for artwork with ID: {}", artwork.getId());
         } else {
             throw new EntityNotFoundException("Artwork not found with ID: " + artwork.getId());
         }
     }
+
 
     public List<Artwork> getArtworksByTitle(String title) {
         Specification<Artwork> titleSpecification = (root, query, criteriaBuilder) ->
